fix: Manual clippy fixes
