+++ src/analyse.rs
@@ -16,89 +16,88 @@ pub trait Rule {
 
 pub struct Analyse {
     rules: HashMap<String, Box<dyn Rule>>,
-    file: File,
 }
 
-use crate::rules::{self, E0010};
+use crate::rules::{self};
 impl Analyse {
-    pub fn new(disable: Vec<String>, file: File) -> Self {
+    pub fn new(disable: Vec<String>) -> Self {
         let mut rules = HashMap::new();
         // @todo refactor this code below
         if !disable.contains(&"E001".to_string()) {
             rules.insert(
                 "E001".to_string(),
-                Box::new(rules::E001::E001 {}) as Box<dyn Rule>,
+                Box::new(rules::e001::E001 {}) as Box<dyn Rule>,
             );
         }
 
         if !disable.contains(&"E002".to_string()) {
             rules.insert(
                 "E002".to_string(),
-                Box::new(rules::E002::E002 {}) as Box<dyn Rule>,
+                Box::new(rules::e002::E002 {}) as Box<dyn Rule>,
             );
         }
 
         if !disable.contains(&"E003".to_string()) {
             rules.insert(
                 "E003".to_string(),
-                Box::new(rules::E003::E003 {}) as Box<dyn Rule>,
+                Box::new(rules::e003::E003 {}) as Box<dyn Rule>,
             );
         }
 
         if !disable.contains(&"E004".to_string()) {
             rules.insert(
                 "E004".to_string(),
-                Box::new(rules::E004::E004 {}) as Box<dyn Rule>,
+                Box::new(rules::e004::E004 {}) as Box<dyn Rule>,
             );
         }
         if !disable.contains(&"E005".to_string()) {
             rules.insert(
                 "E005".to_string(),
-                Box::new(rules::E005::E005 {}) as Box<dyn Rule>,
+                Box::new(rules::e005::E005 {}) as Box<dyn Rule>,
             );
         }
 
         if !disable.contains(&"E006".to_string()) {
             rules.insert(
                 "E006".to_string(),
-                Box::new(rules::E006::E006 {}) as Box<dyn Rule>,
+                Box::new(rules::e006::E006 {}) as Box<dyn Rule>,
             );
         }
 
         if !disable.contains(&"E007".to_string()) {
             rules.insert(
                 "E007".to_string(),
-                Box::new(rules::E007::E007 {}) as Box<dyn Rule>,
+                Box::new(rules::e007::E007 {}) as Box<dyn Rule>,
             );
         }
 
         if !disable.contains(&"E008".to_string()) {
             rules.insert(
                 "E008".to_string(),
-                Box::new(rules::E008::E008 {}) as Box<dyn Rule>,
+                Box::new(rules::e008::E008 {}) as Box<dyn Rule>,
             );
         }
 
         if !disable.contains(&"E009".to_string()) {
             rules.insert(
                 "E009".to_string(),
-                Box::new(rules::E009::E009 {}) as Box<dyn Rule>,
+                Box::new(rules::e009::E009 {}) as Box<dyn Rule>,
             );
         }
         if !disable.contains(&"E0010".to_string()) {
             rules.insert(
                 "E0010".to_string(),
-                Box::new(E0010::E0010::new(file.clone())) as Box<dyn Rule>,
+                Box::new(rules::e0010::E0010 {}) as Box<dyn Rule>,
             );
         }
         if !disable.contains(&"E0011".to_string()) {
             rules.insert(
                 "E0011".to_string(),
-                Box::new(rules::E0011::E0011 {}) as Box<dyn Rule>,
+                Box::new(rules::e0011::E0011 {}) as Box<dyn Rule>,
             );
         }
         
-        Self { rules, file }
+        Self { rules }
     }
 
     pub fn statement(&self, statement: parser::ast::Statement) -> Vec<Suggestion> {
@@ -110,24 +109,23 @@ impl Analyse {
         suggestions
     }
 
+    #[allow(clippy::only_used_in_recursion)]
+    #[allow(clippy::borrowed_box)]
     fn expand(&self, statement: &Statement, rule: &Box<dyn Rule>) -> Vec<Suggestion> {
         let mut suggestions = Vec::new();
         suggestions.append(&mut rule.validate(statement));
         match statement {
             Statement::Try(s) => {
                 for catch in &s.catches {
-                    match catch {
-                        CatchBlock {
-                            start: _,
-                            end: _,
-                            types: _,
-                            var: _,
-                            body,
-                        } => {
-                            for statement in body {
-                                suggestions.append(&mut self.expand(statement, rule));
-                            }
-                        }
+                    let CatchBlock {
+                        start: _,
+                        end: _,
+                        types: _,
+                        var: _,
+                        body,
+                    } = catch;
+                    for statement in body {
+                        suggestions.append(&mut self.expand(statement, rule));
                     }
                 }
             }
@@ -164,14 +162,15 @@ impl Analyse {
                     };
                 }
             }
-            Statement::If(if_statement) => match if_statement {
-                IfStatement {
+            Statement::If(if_statement) => {
+                let IfStatement {
                     r#if: _,
                     left_parenthesis: _,
                     condition: _,
                     right_parenthesis: _,
                     body,
-                } => {
+                } = if_statement;
+                {
                     match body {
                         IfStatementBody::Block {
                             colon: _,
