+++ src/rules/e005.rs
@@ -7,18 +7,14 @@ pub struct E005 {}
 impl Rule for E005 {
     fn validate(&self, statement: &Statement) -> Vec<Suggestion> {
         let mut suggestions = Vec::new();
-        match statement {
-            Statement::Class(class) => {
-                let name = String::from(class.name.value.clone());
-                match has_capitalized_name(name.clone(), class.class) {
-                    Some(s) => {
-                        suggestions.push(s);
-                    }
-                    None => {}
-                }
+
+        if let Statement::Class(class) = statement {
+            let name = String::from(class.name.value.clone());
+            if let Some(s) = has_capitalized_name(name.clone(), class.class) {
+                suggestions.push(s);
             }
-            _ => {}
-        }
+        };
+
         suggestions
     }
 }
