+++ src/rules/mod.rs
@@ -1,11 +1,11 @@
-pub mod E001;
-pub mod E0010;
-pub mod E0011;
-pub mod E002;
-pub mod E003;
-pub mod E004;
-pub mod E005;
-pub mod E006;
-pub mod E007;
-pub mod E008;
-pub mod E009;
+pub mod e001;
+pub mod e002;
+pub mod e003;
+pub mod e004;
+pub mod e005;
+pub mod e006;
+pub mod e007;
+pub mod e008;
+pub mod e009;
+pub mod e0010;
+pub mod e0011;
